requirements.txt
requests==2.31.0
twilio==8.2.0
python-dotenv==1.0.0